github-bot/requirements.txt
Everett==1.0.2
Flask==1.1.2
PyGitHub==1.51
